feature2d: rework opencl perf tests
use the same configuration as cpu tests
